update feature to allow parent declaration the health profile for the…

+++ SWP_BACKEND/SchoolMedicalManagementSystem/SchoolMedicalManagementSystem/src/main/java/group6/Swp391/Se1861/SchoolMedicalManagementSystem/model/ItemRequest.java
@@ -28,7 +28,7 @@ public class ItemRequest {
     private String itemType;
 
     @Column(name = "dosage", nullable = false)
-    private int dosage;
+    private double dosage;
 
     @Column(name = "frequency", nullable = false)
     private int frequency;
